Moved classes to new clustering tests packages
