--- src/test/java/io/vertx/test/core/ZKClusteredComplexHATest.java
@@ -1,22 +1,0 @@
-package io.vertx.test.core;
-
-import io.vertx.core.spi.cluster.ClusterManager;
-import io.vertx.spi.cluster.zookeeper.MockZKCluster;
-
-/**
- * Created by stream.Liu
- */
-public class ZKClusteredComplexHATest extends ComplexHATest {
-
-  private MockZKCluster zkClustered = new MockZKCluster();
-
-  public void after() throws Exception {
-    super.after();
-    zkClustered.stop();
-  }
-
-  @Override
-  protected ClusterManager getClusterManager() {
-    return zkClustered.getClusterManager();
-  }
-}
